// frontend/src/app/core/api/model/exchange-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CountryResponse } from './country-response';


export interface ExchangeResponse { 
    id?: string;
    mic?: string;
    symbol?: string;
    name?: string;
    country?: CountryResponse;
}