// frontend/src/app/core/api/model/exchange-add-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExchangeAddRequest { 
    name: string;
    mic: string;
    acronym: string;
    countryId: string;
}